move hedgehog_algorithm into abstract
