+++ src/hhimg/abstract/hedgehog_algorithm.h
